+++ src/shapes/image.class.ts
@@ -17,6 +17,7 @@ import {
   LoadImageOptions,
 } from '../util/misc/objectEnlive';
 import { parsePreserveAspectRatioAttribute } from '../util/misc/svgParsing';
+import { classRegistry } from '../util/class_registry';
 import { FabricObject, fabricObjectDefaultValues } from './Object/FabricObject';
 
 export type ImageSource =
@@ -722,15 +723,13 @@ export class Image extends FabricObject {
    * @returns {Promise<Image>}
    */
   static fromObject(
-    { filters: f, resizeFilter, src, crossOrigin, ...object }: any,
+    { filters: f, resizeFilter: rf, src, crossOrigin, ...object }: any,
     options: { signal: AbortSignal }
   ): Promise<Image> {
-    const imageOptions = { ...options, crossOrigin },
-      filterOptions = { ...options, namespace: filters };
     return Promise.all([
-      loadImage(src, imageOptions),
-      f && enlivenObjects(f, filterOptions),
-      resizeFilter && enlivenObjects([resizeFilter], filterOptions),
+      loadImage(src, { ...options, crossOrigin }),
+      f && enlivenObjects(f, options),
+      rf && enlivenObjects([rf], options),
       enlivenObjectEnlivables(object, options),
     ]).then(([el, filters = [], [resizeFilter] = [], hydratedProps = {}]) => {
       return new this(el, {
@@ -795,5 +794,8 @@ export const imageDefaultValues: Partial<TClassProperties<Image>> = {
 };
 
 Object.assign(Image.prototype, imageDefaultValues);
+
+classRegistry.setClass(Image);
+classRegistry.setSVGClass(Image);
 
 fabric.Image = Image;