+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## [next]
 
-- refactor(): `getRegularPolygonPath` => `Polygon.getRegularPolygonPoints`, `Polygon.createRegularPolygon` [#8537](https://github.com/fabricjs/fabric.js/pull/8537)
+- feat(): `getRegularPolygonPath` => `Polygon.getRegularPolygonPoints`, `Polygon.createRegularPolygon` [#8537](https://github.com/fabricjs/fabric.js/pull/8537)
+- chore(TS): Convert Canvas events mixin and grouping mixin [#8519](https://github.com/fabricjs/fabric.js/pull/8519)
+- chore(TS): Remove backward compatibility initialize methods [#8525](https://github.com/fabricjs/fabric.js/pull/8525/)
+- chore(TS): replace getKlass utility with a registry that doesn't require full fabricJS to work [#8500](https://github.com/fabricjs/fabric.js/pull/8500)
 - chore(): use context in static constructors [#8522](https://github.com/fabricjs/fabric.js/issues/8522)
 - chore(TS): Convert Canvas class #8510
 - chore(TS): Move object classes #8511